requirements.txt
--extra-index-url=https://download.pytorch.org/whl/cu118/
affine==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.0
docstring-parser==0.15
filelock==3.9.0
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
jsonargparse==4.25.0
kiwisolver==1.4.5
lightning-utilities==0.9.0
lit==15.0.7
MarkupSafe==2.1.2
matplotlib==3.8.0
mercantile==1.2.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.0.9.post0
PyYAML==6.0.1
rasterio==1.3.8
requests==2.28.1
safetensors==0.3.3
six==1.16.0
snuggs==1.4.7
sympy==1.12
timm==0.9.7
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.2.0
torchvision==0.15.2+cu118
tqdm==4.66.1
triton==2.0.0
typeshed-client==2.3.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.9.2
